+++ requirements.txt
@@ -6,7 +6,6 @@ requests==2.32.4
 pytest==6.2.5
 black==24.3.0
 flake8==4.0.1
-dulwich==0.20.24
 zipp>=3.19.1 # not directly required, pinned by Snyk to avoid a vulnerability
 setuptools>=78.1.1 # not directly required, pinned by Snyk to avoid a vulnerability
 urllib3>=2.5.0 # not directly required, pinned by Snyk to avoid a vulnerability
