+++ tests/test_autodeploy.py
@@ -1,0 +1,107 @@
+import os
+import sys
+
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
+import unittest
+from AutoDeploy import (
+    AutomateDeployment,
+    remove_dead_code,
+    simplify_expressions,
+    use_efficient_data_structures,
+    minimize_io_operations,
+    utilize_builtin_functions,
+    employ_caching_techniques,
+    profile_and_benchmark,
+)
+
+
+class TestCodeOptimization(unittest.TestCase):
+    def test_remove_dead_code(self):
+        code = '''def remove_dead_code():
+    print("This is dead code")
+
+def test_function():
+    print("This is a test function")'''
+        expected_code = '''def test_function():
+    print("This is a test function")'''
+        cleaned_code = remove_dead_code(code)
+        self.assertEqual(cleaned_code.strip(), expected_code.strip())
+
+    def test_simplify_expressions(self):
+        code = '''x = 1 + 1
+y = 2 * 5'''
+        expected_code = '''x = 2
+y = 10'''
+        simplified_code = simplify_expressions(code)
+        self.assertEqual(simplified_code.strip(), expected_code.strip())
+
+    def test_use_efficient_data_structures(self):
+        code = '''data = list(set([1, 2, 3, 3, 4, 5]))'''
+        expected_code = '''data = set([1, 2, 3, 4, 5])'''
+        optimized_code = use_efficient_data_structures(code)
+        self.assertEqual(optimized_code.strip(), expected_code.strip())
+
+    def test_minimize_io_operations(self):
+        code = '''data = open(file)
+# Perform operations on contents'''
+        expected_code = '''data = cached_file
+# Perform operations on contents'''
+        optimized_code = minimize_io_operations(code)
+        self.assertEqual(optimized_code.strip(), expected_code.strip())
+
+    def test_utilize_builtin_functions(self):
+        code = '''custom_function()'''
+        expected_code = '''built_in_function()'''
+        optimized_code = utilize_builtin_functions(code)
+        self.assertEqual(optimized_code.strip(), expected_code.strip())
+
+    def test_employ_caching_techniques(self):
+        code = '''result = compute_expensive_operation()'''
+        expected_code = '''result = cached_result'''
+        optimized_code = employ_caching_techniques(code)
+        self.assertEqual(optimized_code.strip(), expected_code.strip())
+
+    def test_profile_and_benchmark(self):
+        code = '''# Critical section of code
+for i in range(1000000):
+    pass'''
+        expected_code = code
+        profiled_code = profile_and_benchmark(code)
+        self.assertEqual(profiled_code.strip(), expected_code.strip())
+
+
+class TestCode(unittest.TestCase):
+    def test_deploy(self):
+        combined_library = 'my_combined_library'
+        deployment = AutomateDeployment(combined_library)
+        self.assertIsNone(deployment.deploy())
+
+    def test_get_repos_libraries(self):
+        combined_library = 'my_combined_library'
+        deployment = AutomateDeployment(combined_library)
+        repos, libraries = deployment.get_repos_libraries()
+        self.assertIsNotNone(repos)
+        self.assertIsNotNone(libraries)
+
+    def test_auto_credit(self):
+        combined_library = 'my_combined_library'
+        deployment = AutomateDeployment(combined_library)
+        repos = ['repo1', 'repo2']
+        libraries = ['library1', 'library2']
+        self.assertIsNone(deployment.auto_credit(repos, libraries))
+
+    def test_fork_and_comment(self):
+        combined_library = 'my_combined_library'
+        deployment = AutomateDeployment(combined_library)
+        repo_data = {'full_name': 'my_repo'}
+        self.assertIsNone(deployment.fork_and_comment(repo_data))
+
+    def test_log_library_usage(self):
+        combined_library = 'my_combined_library'
+        deployment = AutomateDeployment(combined_library)
+        library_name = 'my_library'
+        self.assertIsNone(deployment.log_library_usage(library_name))
+
+
+if __name__ == '__main__':
+    unittest.main()
